:recycle: major refactor
